Added custom orbital elements input
Added functionality to input and run Lambert solver on comet with custom orbital elements.

KNOWN ISSUE:
Custom orbit is currently propagated in discrete time steps. If the eccentricity gets too close to 1, then the orbit error is so large that it is unusable. If e > 1.1 then all is good. Will be reworking logic to hopefully fix this, since there exist comets like C/2024 E1 with eccentricities of 1.000055